vector-store: move Alternator ":attrs" decoding out of vector.rs
parse_alternator_scalar() has nothing to do with vectors - it decodes
a filtering/partition-key attribute's ":attrs" value into a scalar
CqlValue (Text/Blob/Decimal), for the mechanism added a few commits up
to let Alternator declare types for such columns. It only ended up in
vector.rs because it happens to share Alternator's tagged-blob wire
format with the embedding-vector decoding that already lived there.

Move all of Alternator's ":attrs" wire-format decoding - the
ALTERNATOR_TYPE_* tag constants, parse_alternator_scalar(), and the
vector-specific parse_alternator_vector()/parse_alternator_vector_binary()/
parse_alternator_list_json() - into a new alternator.rs module. The tag
constants are one cohesive group mirroring a single C++ enum
(alternator::alternator_type in alternator/serialization.hh, per the
existing "do not reorder" comment), so splitting them across two Rust
modules by vector-vs-scalar consumer would obscure that they must stay
in sync as a unit.

vector.rs now only has the Vector type and its TryFrom<CqlValue> impl,
which calls into alternator::parse_alternator_vector() for the
CqlValue::Blob case exactly as it calls the CQL driver's own
deserialization for CqlValue::Vector - both are just "how do I get a
Vector out of this CqlValue" branches, with no Alternator-specific
knowledge leaking into vector.rs itself.

Pure refactor, no behavioral change: parse_alternator_scalar()'s and
parse_alternator_vector()'s unit tests moved with them; Vector's own
tests (which exercise the Alternator-encoded-blob path indirectly
through TryFrom) stay in vector.rs, since they're testing Vector's
public contract, not alternator.rs's internals.
